Move back to ballerina-cli, show error code
