+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/utils/DiagnosticAnnotation.java
@@ -1,4 +1,4 @@
-package io.ballerina.shell.cli;
+package io.ballerina.cli.utils;
 
 import io.ballerina.tools.diagnostics.DiagnosticSeverity;
